refactor(pragma): split tracker into srp modules (#0000)
